refactor(sema): combine NatSpec outputs
