+++ crates/sema/src/output/mod.rs
@@ -1,17 +1,16 @@
 use serde::Serialize;
 
 mod abi;
-mod devdoc;
+mod natspec;
 mod storage_layout;
-mod userdoc;
 
-pub use devdoc::{DevDocItem, DevDocumentation, StateVariableDoc};
+pub use natspec::{
+    DevDocItem, DevDocumentation, StateVariableDoc, UserDocNotice, UserDocumentation,
+};
 pub use storage_layout::{
     StorageEncoding, StorageLayoutEntry, StorageLayoutMember, StorageLayoutOutput,
     StorageLayoutType,
 };
-pub use userdoc::{UserDocNotice, UserDocumentation};
-
 #[derive(Debug, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum DocumentationKind {
